# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataForSEO Intelligence API
  version: 1.0.0
  summary: Multi-tenant SEO research API on top of DataForSEO v3.
  description: 'A multi-tenant SaaS SEO platform. Every billed call flows through a three-tier cost engine
    (Redis -> Postgres -> upstream) with singleflight coalescing and stale-while-revalidate, so repeat
    reads are $0.


    **Conventions**

    - All endpoints are versioned under `/api/v1`.

    - Every data response embeds a `meta` object (`from_cache`, `cost_cents`, `source`, `latency_ms`).

    - All money values are integer **USD cents**.

    - Auth is JWT Bearer; access tokens are short-lived, refreshed via `/auth/refresh`.

    - Per-organization monthly quota is enforced before any billed call (HTTP 402).'
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
paths:
...
  /api/v1/rank/track:
    post:
      tags:
      - rank
      summary: Track
      operationId: track_api_v1_rank_track_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RankTrackRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankTrackResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/rank/history:
    get:
      tags:
      - rank
      summary: History
      operationId: history_api_v1_rank_history_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: keyword
        in: query
        required: true
        schema:
          type: string
          title: Keyword
      - name: domain
        in: query
        required: true
        schema:
          type: string
          title: Domain
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankHistoryResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/rank/tracked:
    get:
      tags:
      - rank
      summary: Tracked
      operationId: tracked_api_v1_rank_tracked_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackedListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/schedules:
    get:
      tags:
      - schedules
      summary: List Schedules
      operationId: list_schedules_api_v1_schedules_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - schedules
      summary: Create Schedule
      operationId: create_schedule_api_v1_schedules_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/schedules/{schedule_id}:
    patch:
      tags:
      - schedules
      summary: Update Schedule
      operationId: update_schedule_api_v1_schedules__schedule_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          title: Schedule Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - schedules
      summary: Delete Schedule
      operationId: delete_schedule_api_v1_schedules__schedule_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          title: Schedule Id
      responses:
        '204':
          description: Successful Response
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/schedules/{schedule_id}/run:
    post:
      tags:
      - schedules
      summary: Run Now
      description: Trigger a schedule immediately (also used by external cron).
      operationId: run_now_api_v1_schedules__schedule_id__run_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          title: Schedule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Meta:
      properties:
        from_cache:
          type: boolean
          title: From Cache
        cost_cents:
          type: integer
          title: Cost Cents
        source:
          type: string
          title: Source
        latency_ms:
          type: integer
          title: Latency Ms
      type: object
      required:
      - from_cache
      - cost_cents
      - source
      - latency_ms
      title: Meta
...
    RankHistoryResponse:
      properties:
        keyword:
          type: string
          title: Keyword
        domain:
          type: string
          title: Domain
        history:
          items:
            $ref: '#/components/schemas/RankPoint'
          type: array
          title: History
          default: []
      type: object
      required:
      - keyword
      - domain
      title: RankHistoryResponse
    RankPoint:
      properties:
        position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Position
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - created_at
      title: RankPoint
    RankTrackRequest:
      properties:
        keyword:
          type: string
          maxLength: 255
          minLength: 1
          title: Keyword
        domain:
          type: string
          maxLength: 255
          minLength: 1
          title: Domain
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        depth:
          type: integer
          maximum: 100.0
          minimum: 10.0
          title: Depth
          default: 100
      type: object
      required:
      - keyword
      - domain
      title: RankTrackRequest
    RankTrackResponse:
      properties:
        keyword:
          type: string
          title: Keyword
        domain:
          type: string
          title: Domain
        position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Position
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        found:
          type: boolean
          title: Found
        depth:
          type: integer
          title: Depth
        history:
          items:
            $ref: '#/components/schemas/RankPoint'
          type: array
          title: History
          default: []
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - keyword
      - domain
      - found
      - depth
      - meta
      title: RankTrackResponse
...
    ScheduleCreate:
      properties:
        kind:
          type: string
          const: site_report
          title: Kind
          default: site_report
        frequency:
          type: string
          enum:
          - daily
          - weekly
          - monthly
          title: Frequency
          default: weekly
        params:
          type: object
          title: Params
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
      type: object
      title: ScheduleCreate
    ScheduleListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ScheduleOut'
          type: array
          title: Items
          default: []
      type: object
      title: ScheduleListResponse
    ScheduleOut:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          title: Kind
        frequency:
          type: string
          title: Frequency
        params:
          type: object
          title: Params
        project_id:
          type: string
          title: Project Id
        active:
          type: boolean
          title: Active
        next_run_at:
          type: string
          title: Next Run At
        last_run_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Run At
        last_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Status
        label:
          type: string
          title: Label
      type: object
      required:
      - id
      - kind
      - frequency
      - params
      - project_id
      - active
      - next_run_at
      - label
      title: ScheduleOut
    ScheduleUpdate:
      properties:
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
        frequency:
          anyOf:
          - type: string
            enum:
            - daily
            - weekly
            - monthly
          - type: 'null'
          title: Frequency
      type: object
      title: ScheduleUpdate
...
    TrackedItem:
      properties:
        keyword:
          type: string
          title: Keyword
        domain:
          type: string
          title: Domain
        location_code:
          type: integer
          title: Location Code
        language_code:
          type: string
          title: Language Code
        latest_position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Latest Position
        previous_position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Previous Position
        delta:
          anyOf:
          - type: integer
          - type: 'null'
          title: Delta
        last_checked:
          type: string
          title: Last Checked
        observations:
          type: integer
          title: Observations
      type: object
      required:
      - keyword
      - domain
      - location_code
      - language_code
      - last_checked
      - observations
      title: TrackedItem
    TrackedListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TrackedItem'
          type: array
          title: Items
          default: []
      type: object
      title: TrackedListResponse
...
    Problem:
      type: object
      description: RFC 7807 Problem Details (media type `application/problem+json`).
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
          format: uri
          examples:
          - https://docs.seo-intelligence.app/errors/quota-exceeded
        title:
          type: string
          examples:
          - Payment Required
        status:
          type: integer
          examples:
          - 402
        detail:
          type: string
          examples:
          - Monthly organization quota exhausted.
        instance:
          type: string
          examples:
          - /api/v1/serp/ranking
        errors:
          type: array
          description: Field-level validation failures (422 only).
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/login` or `/auth/refresh`.
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PaymentRequired:
      description: Monthly organization quota exhausted — top up or wait for the next cycle.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource does not exist or is not visible to you.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Upstream rate limit hit; retry after the indicated delay.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
    ValidationFailed:
      description: Request body or parameters failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.example.com
  description: Production
...
security:
- BearerAuth: []